+++ start/src/App.vue
@@ -1,25 +1,61 @@
 <script lang="ts" setup>
-  import { ref } from 'vue'
+  import { ref, watch, reactive } from 'vue'
 
-  const tags = ref([1, 2, 3, 4, 5, 6, 7, 8])
+  /** 省份列表 */
+  const provinceList = ref()
+  const cityList = ref()
 
-  const close = (i: number): void => {
-    tags.value.splice(i, 1)
+  const form = reactive({
+    province: '',
+    city: ''
+  })
+
+  /** 获取省份列表 */
+  const getProvinceList = (): void => {
+    fetch(
+      'https://restapi.amap.com/v3/config/district?keywords=%E4%B8%AD%E5%9B%BD&subdistrict=1&key=ca2358293d08bfa6028bb8fb1c076130'
+    )
+      .then(res => res.json())
+      .then(data => {
+        provinceList.value = data.districts[0].districts
+        console.log(provinceList.value)
+      })
   }
+
+  getProvinceList()
+
+  watch(
+    () => form.province,
+    (newVal?: string): void => {
+      fetch(
+        `https://restapi.amap.com/v3/config/district?keywords=${newVal}&subdistrict=1&key=ca2358293d08bfa6028bb8fb1c076130`
+      )
+        .then(res => res.json())
+        .then(data => {
+          cityList.value = data.districts[0].districts
+        })
+    }
+  )
 </script>
 
 <template>
   <f-space>
-    <f-tag
-      v-for="(t, i) in tags"
-      :key="i"
-      round
-      type="primary"
-      simple
-      close
-      :on-close="() => close(i)"
-    >
-      {{ t }}
-    </f-tag>
+    <!-- 选择省份 -->
+    <f-select v-if="provinceList" v-model="form.province" :width="200">
+      <f-option
+        v-for="(province, index) in provinceList"
+        :key="index"
+        :value="province.name"
+      >
+        {{ province.name }}
+      </f-option>
+    </f-select>
+
+    <!-- 选择城市 -->
+    <f-select v-if="cityList" v-model="form.city">
+      <f-option v-for="(city, index) in cityList" :key="index" :value="city.name">
+        {{ city.name }}
+      </f-option>
+    </f-select>
   </f-space>
 </template>